requirements.txt.txt
streamlit 
torch 
torchvision 
pillow